% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Driver.R
\name{LoggingDBI}
\alias{LoggingDBI}
\title{LoggingDBI driver}
\usage{
LoggingDBI(drv, logger = get_default_logger())
}
\arguments{
\item{drv}{Driver to be wrapped.}

\item{logger}{Logger object, defaults to \code{\link[=get_default_logger]{get_default_logger()}}.}
}
\description{
TBD.
}
\examples{
\dontrun{
#' library(DBI)
RLoggingDBI::LoggingDBI()
}
}
