% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-log-call.R
\name{get_default_logger}
\alias{get_default_logger}
\alias{make_console_logger}
\alias{make_collect_logger}
\title{Logging parameters}
\usage{
get_default_logger()

make_console_logger()

make_collect_logger()
}
\description{
TBD.
}
